The proxies go to detail/, one header per iterator category (#131)
Issue #131's rule is that a header a user never names lives in detail/, and
the four proxies are the cleanest case in the tree: bit_set_iterator,
bit_set_reference, bit_sequence_iterator and bit_sequence_reference appear
nowhere in test/, benchmark/ or README.md outside the one test source that
mirrored their header. They are reached through a container's iterator and
reference typedefs and through nothing else.

So bit_proxy.hpp splits along the line that already ran through it -- the
set reading's proxy is bidirectional, the sequence reading's is random
access -- into detail/bidirectional.hpp and detail/random_access.hpp. The
two pairs never referenced each other, so the split is exact.

The names follow the headers: bidirectional_bit_iterator,
bidirectional_bit_reference, random_access_bit_iterator and
random_access_bit_reference, in xstd::detail::bits beside the other four
detail headers. The category names the iterator, where the reading names the
container that hands it out; the bit_ infix says what is iterated, as bit_
says what is stored in the container names, and it is what keeps
bidirectional_bit_iterator clear of std::bidirectional_iterator, whose
spelling the category alone would have taken.

The walks stay spelled detail::bits::find_next<Traits>(...) inside
xstd::detail::bits itself. Dropping the now-redundant-looking qualification
would reintroduce exactly the hazard the nesting exists to close: an
unqualified call with an explicit template argument performs ADL, and the
associated namespace of the storage being walked is std or boost.
[design.md#why-nested]

The umbrella stops exporting them, and the two adaptors and format.hpp name
them qualified.

The test tree follows the library tree rather than taking the exception
detail/ is granted: bit_proxy.cpp becomes detail/bidirectional.cpp and
detail/random_access.cpp. Unnameable is not unobservable -- what these types
do is the observable behaviour of every container's iterator -- so the
mirror is worth keeping by hand where the rule stops asking for it. Every
assertion moves across unchanged; the cases that asserted the two proxies
against each other become one case on each side, and the static_assert that
conjoined both iterators' triviality becomes one on each. The rule's comment
in test/CMakeLists.txt now says why detail/ is excepted and what a test
under it is: white-box, rewritable with the internals, while the contract
those types carry stays asserted through the containers in
test/include/test/{set,sequence}/concepts.hpp.

Also updated: the -Wno-lifetime-safety-intra-tu-suggestions note in test/ and
benchmark/, which still named xstd::proxy::bidirectional -- a spelling that
last existed before #98 consolidated those headers.
