+++ CMakeLists.txt
@@ -79,7 +79,6 @@ target_sources(
         include/xstd/bits/set_adaptor.hpp
         include/xstd/bits/bitset_adaptor.hpp
         include/xstd/bits/bit_array.hpp
-        include/xstd/bits/bit_proxy.hpp
         include/xstd/bits/bit_set.hpp
         include/xstd/bits/bit_set_view.hpp
         include/xstd/bits/bit_span.hpp
@@ -90,11 +89,13 @@ target_sources(
         include/xstd/bits/bit_traits.hpp
         include/xstd/bits/block_sequence.hpp
         include/xstd/bits/detail/allocator_typedef.hpp
+        include/xstd/bits/detail/bidirectional.hpp
         include/xstd/bits/detail/hash.hpp
         include/xstd/bits/detail/intrin.hpp
         include/xstd/bits/dynamic_bitset.hpp
         include/xstd/bits/format.hpp
         include/xstd/bits/detail/pred.hpp
+        include/xstd/bits/detail/random_access.hpp
         include/xstd/bits/ext/boost.hpp
         include/xstd/bits/ext/boost/dynamic_bitset.hpp
         include/xstd/bits/ext/std.hpp
