+++ test/CMakeLists.txt
@@ -39,7 +39,7 @@ set(cxx_compile_options_warnings
     $<$<CXX_COMPILER_ID:Clang>:
         -Weverything
         -Wno-unknown-warning-option         # forward compatibility for new Clang warnings
-        -Wno-lifetime-safety-intra-tu-suggestions             # Clang trunk (23-SVN); suggests [[clang::lifetimebound]] across xstd::proxy::bidirectional and xstd::bitset - not yet acted on
+        -Wno-lifetime-safety-intra-tu-suggestions             # Clang trunk (23-SVN); suggests [[clang::lifetimebound]] across xstd::detail::bits::bidirectional_bit_iterator and xstd::bitset - not yet acted on
         -Wno-lifetime-safety-intra-tu-constructor-suggestions # ditto, for constructors
         -Wno-lifetime-safety-invalidation                     # ditto; false positive on a.clear() in test/include/test/set/primitives.hpp treating the container itself as invalidated
         -Wno-c++98-compat
@@ -142,6 +142,10 @@ foreach(h ${public_headers})
 endforeach()
 
 # Every public header is answered by one test source at the same path, detail/ excepted; enforced here so it cannot drift.
+# detail/ is excepted because a detail header has no contract to mirror. Where such a test exists all the same -- and the
+# two proxy headers have one each -- it is white-box: it names what nobody outside the library names, and may be rewritten
+# whenever the internals are. The contract those types do carry is asserted through the containers that hand them out,
+# in test/include/test/set/concepts.hpp and test/include/test/sequence/concepts.hpp.
 file(
     GLOB_RECURSE mirrored_headers
     CONFIGURE_DEPENDS
